Move form auth and basic realm to runtime
